+++ NEWS.md
@@ -1,3 +1,5 @@
+# autodb (development version)
+
 # autodb 3.0.0
 
 ## Breaking changes