+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Improvements
 
 * Changed `autoref` (and `normalise`) to spend less time removing extraneous references.
+* Changed `autodb` (and optionally `database`) to not check foreign key references.
 
 # autodb 3.0.0