+++ NEWS.md
@@ -1,4 +1,4 @@
-# autodb (development version)
+# autodb 3.1.0
 
 ## Improvements